requirements.txt
requests
python-qbittorrent